// tteok-shop/src/libs/types/enums/product.enum.ts
export enum ProductCollection {
  TTEOK = "TTEOK",
  HANGWA = "HANGWA",
  SIKHYE = "SIKHYE",
  SUJEONGGWA = "SUJEONGGWA",
  OMIJA = "OMIJA",
  SET = "SET",
}

export enum ProductSize {
  SMALL = "SMALL",
  NORMAL = "NORMAL",
  LARGE = "LARGE",
  FAMILY = "FAMILY",
}

export enum ProductVolume {
  ML_300 = 300,
  ML_500 = 500,
  ML_750 = 750,
  L_1 = 1000,
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}